Fix songqueue from freezing (for real this time)
Instead of trying to make the threaded solution work, I just reworked the soundplayer to run in a subprocess instead. Now it works.
