+++ plugins/soundreq/soundrequesting.py
@@ -1,14 +1,9 @@
+from .soundplayer import SoundRequestPlayer
 import config
 import events
-import json
 import os
 import plugins
-import requests
 import threading
-import time
-import traceback
-import vlc
-import websocket
 
 queue_lock = threading.Lock()
 queue:list[tuple[str, str|None, str|None]] = []
@@ -73,8 +68,8 @@ def handler_target(sound_keys:list[tuple[str, str|None, str|None]]=None):
             sound_done.clear()
             print("Playing Sound", key)
             events.dispatch(events.Event("soundreq:play_sound", {"success": True, "key":key, "sound": info}))
-
-            sound_done.wait()
+            print("waiting", sound_done.is_set())
+            print("x", sound_done.wait())
             
             print("Stopped Sound", key)
 
@@ -87,99 +82,6 @@ def handler_target(sound_keys:list[tuple[str, str|None, str|None]]=None):
 
 def invoke_handler():
     global queue_handler
-    print("DEBUG:", queue_handler)
     if queue_handler is None:
         queue_handler = threading.Thread(target=handler_target, args=(popall_queue(),), daemon=True)
         queue_handler.start()
-
-
-class SoundRequestPlayer:
-    def __init__(self, api_url_host:str, api_secure:bool=False):
-        self.vlc_instance:vlc.Instance = None
-        self.vlc_player:vlc.MediaPlayer = None
-        self.api_url_host = api_url_host
-        self.api_secure = api_secure
-        self.wsa:websocket.WebSocketApp = None
-        self.listeners = events.EventListenerCollection({
-            "soundreq:play_sound": [events.EventListener(self.on_play_sound)]
-        })
-    
-    def load_sound(self, key:str):
-        s = "s"*self.api_secure
-        sound:vlc.Media = self.vlc_instance.media_new(f"http{s}://{self.api_url_host}/api/soundreq/sound/{key}")
-        return sound
-    
-    def on_play_sound(self, event:events.Event):
-        key = event.data.get("key",None)
-        if key is not None and event.data.get("success",True):
-            sound = self.load_sound(key)
-            self.vlc_player.set_media(sound)
-            time.sleep(0.1)
-            self.vlc_player.play()
-
-    def init_vlc(self):
-        if self.vlc_instance is None:
-            self.vlc_instance = vlc.Instance("--input-repeat=-1", "--fullscreen", "--file-caching=0")
-        if self.vlc_player is None:
-            self.vlc_player = self.vlc_instance.media_player_new()
-
-    #cite: https://stackoverflow.com/a/73886462
-    def get_device(self, name:str):
-        mods = self.vlc_player.audio_output_device_enum()
-        if mods:
-            mod = mods
-            while mod:
-                mod = mod.contents
-                if name in str(mod.description):
-                    vlc.libvlc_audio_output_device_list_release(mods)
-                    return mod.device, mod.description
-                mod = mod.next
-        vlc.libvlc_audio_output_device_list_release(mods)
-        return None, None
-    
-    def _end_reached(self, _):
-        self.vlc_player.set_media(None)
-        s = "s"*self.api_secure
-        requests.post(f"http{s}://{self.api_url_host}/api/soundreq/end")
-    
-    def ws_on_open(self, ws:websocket.WebSocket):
-        self.init_vlc()
-        configs_parent = get_configs()
-        configs:dict = configs_parent.get("Sound-Request", {})
-        if "Output-Device" in configs:
-            device, _ = self.get_device(configs["Output-Device"])
-            vlc.libvlc_audio_output_device_set(self.vlc_player, None, device)
-        event_manager = self.vlc_player.event_manager()
-        event_manager.event_attach(vlc.EventType.MediaPlayerEndReached, self._end_reached)
-
-    def ws_on_close(self, ws:websocket.WebSocket, status_code:int, msg:str|bytearray|memoryview):
-        print(f"Sound request player connection closed ({status_code}):", msg)
-
-    def ws_on_message(self, ws:websocket.WebSocket, msg:str|bytearray|memoryview):
-        if isinstance(msg, memoryview):
-            msg = msg.tobytes()
-        data = json.loads(msg)
-        event = events.Event(**data)
-        self.listeners.handle_event(event)
-
-    def ws_on_error(self, ws:websocket.WebSocket, e:Exception):
-        traceback.print_exception(e)
-
-    def start(self):
-        s = "s"*self.api_secure
-        self.wsa = websocket.WebSocketApp(f"ws{s}://{self.api_url_host}/api/events", on_open=self.ws_on_open, on_close=self.ws_on_close, on_message=self.ws_on_message, on_error=self.ws_on_error)
-        try:
-            self.wsa.run_forever()
-        except KeyboardInterrupt:
-            pass
-
-    def end(self):
-        if self.wsa is not None:
-            if self.wsa.keep_running:
-                self.wsa.close()
-            self.wsa = None
-        if self.vlc_player is not None:
-            if self.vlc_player.is_playing():
-                self.vlc_player.pause()
-            self.vlc_player.set_media(None)
-            self.vlc_player = None
